Adding working CLI and API scripts
This API allows you to pull down the entire game catalog (or a subset)
of a user on Backloggery. It returns only the title and console (with
original console in there as that string). Needs some work but it's
enough to build a simple cli to check and play with the tool