Added support for streams and new Game props
You can now succcessfully stream the HTML to the CLI provided instead of
having it make HTTP requests for you. This allows more control from the
shell for how the input is handled.

The API now supports returning the rating, completion state (is completed
or not) as well as a field for the original system split from the system
added.